+++ src/flash/nor/dwcssi/dwcssi.h
@@ -202,7 +202,7 @@ void dwcssi_config_trans(struct flash_bank *bank, struct dwcssi_trans_config *tr
 int dwcssi_tx(struct flash_bank *bank, uint32_t in);
 int dwcssi_txwm_wait(struct flash_bank* bank);
 int dwcssi_flash_tx_cmd(struct flash_bank *bank, uint8_t *cmd, uint8_t len, uint8_t cmd_mode);
-int dwcssi_rd_flash_reg(struct flash_bank *bank, uint32_t* rd_val, uint8_t cmd, uint32_t len);
+int dwcssi_rd_flash_reg(struct flash_bank *bank, uint8_t* rd_val, uint8_t cmd, uint32_t len);
 int dwcssi_wr_flash_reg(struct flash_bank *bank, uint8_t *cmd, uint8_t len, uint8_t cmd_mode);
 
 //general spi ops
